# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cohort — AI agent fleet governance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List agents
      parameters:
        - name: platform
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [observation, active, flagged, retiring, retired]
        - name: verdict
          in: query
          required: false
          schema:
            type: string
            enum: [promote, mentor, retire, observation]
        - name: severity
          in: query
          required: false
          schema:
            type: string
            enum: [critical, high, medium, stable]
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Admit a new agent (Carteira de Trabalho)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInput"
      responses:
        "201":
          description: Agent admitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetail"
  /discovery/analyze:
    post:
      operationId: analyzeAgentSource
      tags: [agents]
      summary: Analyze pasted agent code and skills into an editable draft
      description: >-
        Uses an AI model to read the agent's code and skill definitions and
        propose a persona, Carteira de Trabalho and 5-layer KPI metrics. The
        result is an editable draft and is not persisted until admission.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeSourceInput"
      responses:
        "200":
          description: Draft generated from the provided source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDraft"
  /discovery/fetch:
    post:
      operationId: fetchAgentSource
      tags: [agents]
      summary: Fetch agent source from a Git repository or web URL
      description: >-
        Fetches and concatenates the relevant source and skill files from a
        Git repository (e.g. GitHub) or a single web/raw URL, applying size and
        file-type guardrails. Private GitHub repositories are supported when a
        GitHub credential is available through GITHUB_TOKEN, GH_TOKEN, or
        GITHUB_ACCESS_TOKEN; otherwise only public repos are accessible. The
        concatenated text can then be passed to the analyze endpoint. Nothing is
        persisted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FetchSourceInput"
      responses:
        "200":
          description: Source fetched and concatenated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FetchSourceResult"
...
  /agents/{agentId}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get a single agent with full work record
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateAgent
      tags: [agents]
      summary: Update agent core fields
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetail"
    delete:
      operationId: deleteAgent
      tags: [agents]
      summary: Retire and remove an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /agents/{agentId}/identity:
    patch:
      operationId: updateAgentIdentity
      tags: [agents]
      summary: Update the agent work record (identity)
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdentityUpdate"
      responses:
        "200":
          description: Updated agent detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetail"
  /agents/{agentId}/evaluations:
    get:
      operationId: listAgentEvaluations
      tags: [agents]
      summary: Evaluation history for an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Evaluation history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Evaluation"
...
  /agents/{agentId}/verdicts:
    get:
      operationId: listAgentVerdicts
      tags: [agents]
      summary: Immutable verdict history for an agent
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Verdict history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Verdict"
  /agents/{agentId}/verdict/decision:
    post:
      operationId: decideVerdict
      tags: [agents]
      summary: Approve, disagree with, or export the current verdict
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerdictDecisionInput"
      responses:
        "200":
          description: Updated verdict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Verdict"
  /agents/{agentId}/evaluation/metric:
    patch:
      operationId: updateEvaluationMetric
      tags: [agents]
      summary: Edit a KPI metric's goal (target + rationale) on the latest evaluation
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluationMetricUpdate"
      responses:
        "200":
          description: Updated agent detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetail"
...
components:
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Agent:
      type: object
      required:
        - id
        - name
        - slug
        - role
        - platform
        - version
        - status
        - bio
        - currentVerdict
        - verdictConfidence
        - severity
        - healthScore
        - admittedAt
        - lastEvaluatedAt
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        role:
          type: string
        platform:
          type: string
        version:
          type: string
        status:
          type: string
          enum: [observation, active, flagged, retiring, retired]
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: string
        tagline:
          type: string
        monthlyVolume:
          type: integer
        headlineKpis:
          type: array
          items:
            $ref: "#/components/schemas/KpiMetric"
        targetMetrics:
          type: array
          description: >-
            Flattened metrics from the agent's latest evaluation that carry a
            comparable goal. Used by the fleet view to roll up how many metrics
            are currently off-target.
          items:
            $ref: "#/components/schemas/KpiMetric"
        currentVerdict:
          type: string
          enum: [promote, mentor, retire, observation]
        verdictConfidence:
          type: number
        severity:
          type: string
          enum: [critical, high, medium, stable]
        healthScore:
          type: number
        activeAlerts:
          type: integer
        monthlyValue:
          type: number
        monthlyCost:
          type: number
        admittedAt:
          type: string
        lastEvaluatedAt:
          type: string
    AgentDetail:
      type: object
      required:
        - agent
        - identity
        - owners
        - latestEvaluation
      properties:
        agent:
          $ref: "#/components/schemas/Agent"
        identity:
          $ref: "#/components/schemas/AgentIdentity"
        owners:
          $ref: "#/components/schemas/AgentOwners"
        latestEvaluation:
          $ref: "#/components/schemas/Evaluation"
        currentVerdict:
          $ref: "#/components/schemas/Verdict"
    AgentIdentity:
      type: object
      required:
        - bio
        - shouldDo
        - shouldNotDo
        - autonomyLevel
        - limits
        - businessCase
        - version
      properties:
        bio:
          type: string
        shouldDo:
          type: array
          items:
            type: string
        shouldNotDo:
          type: array
          items:
            type: string
        autonomyLevel:
          type: string
          enum: [autonomous, escalates, restricted]
        autonomyNotes:
          type: string
        limits:
          type: array
          items:
            type: string
        businessCase:
          $ref: "#/components/schemas/BusinessCase"
        version:
          type: integer
    BusinessCase:
      type: object
      required:
        - baseline
        - targetPayback
        - actualPayback
        - description
      properties:
        baseline:
          type: string
        targetPayback:
          type: string
        actualPayback:
          type: string
        description:
          type: string
    AgentOwners:
      type: object
      required:
        - businessOwner
        - technicalOwner
        - governanceSponsor
      properties:
        businessOwner:
          type: string
        technicalOwner:
          type: string
        governanceSponsor:
          type: string
    KpiMetric:
      type: object
      required:
        - label
        - value
        - unit
        - trend
      properties:
        label:
          type: string
        value:
          type: number
        unit:
          type: string
        trend:
          type: number
        direction:
          type: string
          enum: [up, down, flat]
        target:
          type: string
        rationale:
          type: string
    KpiLayer:
      type: object
      required:
        - key
        - label
        - score
        - severity
        - metrics
      properties:
        key:
          type: string
          enum: [efficacy, efficiency, adoption, governance, value]
        label:
          type: string
        score:
          type: number
        severity:
          type: string
          enum: [critical, high, medium, stable]
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/KpiMetric"
    EvaluationMetricUpdate:
      type: object
      required:
        - layerKey
        - metricLabel
      properties:
        layerKey:
          type: string
          enum: [efficacy, efficiency, adoption, governance, value]
        metricLabel:
          type: string
        target:
          type: ["string", "null"]
        rationale:
          type: ["string", "null"]
    Evaluation:
      type: object
      required:
        - id
        - agentId
        - evaluatedAt
        - window
        - layers
        - verdict
        - verdictConfidence
        - rationale
      properties:
        id:
          type: string
        agentId:
          type: string
        evaluatedAt:
          type: string
        window:
          type: string
        layers:
          type: array
          items:
            $ref: "#/components/schemas/KpiLayer"
        verdict:
          type: string
          enum: [promote, mentor, retire, observation]
        verdictConfidence:
          type: number
        rationale:
          type: string
...
    NextAction:
      type: object
      required:
        - action
        - owner
        - due
      properties:
        action:
          type: string
        owner:
          type: string
        due:
          type: string
    Verdict:
      type: object
      required:
        - id
        - agentId
        - verdict
        - confidence
        - executionWindow
        - suggestedSponsor
        - nextActions
        - rationale
        - decision
        - createdAt
      properties:
        id:
          type: string
        agentId:
          type: string
        verdict:
          type: string
          enum: [promote, mentor, retire, observation]
        confidence:
          type: number
        executionWindow:
          type: string
        suggestedSponsor:
          type: string
        nextActions:
          type: array
          items:
            $ref: "#/components/schemas/NextAction"
        rationale:
          type: string
        decision:
          type: string
          enum: [pending, approved, disagreed, exported]
        decidedBy:
          type: ["string", "null"]
        decidedAt:
          type: ["string", "null"]
        createdAt:
          type: string
...
    AgentInput:
      type: object
      required:
        - name
        - role
        - platform
        - bio
      properties:
        name:
          type: string
          minLength: 1
        role:
          type: string
          minLength: 1
        platform:
          type: string
        version:
          type: string
        bio:
          type: string
        tagline:
          type: string
        shouldDo:
          type: array
          items:
            type: string
        shouldNotDo:
          type: array
          items:
            type: string
        autonomyLevel:
          type: string
          enum: [autonomous, escalates, restricted]
        autonomyNotes:
          type: string
        limits:
          type: array
          items:
            type: string
        businessOwner:
          type: string
        technicalOwner:
          type: string
        governanceSponsor:
          type: string
        baseline:
          type: string
        targetPayback:
          type: string
        businessCaseDescription:
          type: string
        proposedMetrics:
          type: array
          items:
            $ref: "#/components/schemas/DraftMetric"
    AgentUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        version:
          type: string
        status:
          type: string
          enum: [observation, active, flagged, retiring, retired]
        bio:
          type: string
    IdentityUpdate:
      type: object
      properties:
        bio:
          type: string
        shouldDo:
          type: array
          items:
            type: string
        shouldNotDo:
          type: array
          items:
            type: string
        autonomyLevel:
          type: string
          enum: [autonomous, escalates, restricted]
        autonomyNotes:
          type: string
        limits:
          type: array
          items:
            type: string
        businessOwner:
          type: string
        technicalOwner:
          type: string
        governanceSponsor:
          type: string
    VerdictDecisionInput:
      type: object
      required:
        - decision
      properties:
        decision:
          type: string
          enum: [approved, disagreed, exported]
        reason:
          type: string
        decidedBy:
          type: string
    DraftMetric:
      type: object
      required:
        - layer
        - label
        - unit
        - target
      properties:
        layer:
          type: string
          enum: [efficacy, efficiency, adoption, governance, value]
        label:
          type: string
        unit:
          type: string
        target:
          type: string
        value:
          type: number
          description: >-
            Optional reviewer-set starting/current value for the metric. When
            provided during admission it overrides the deterministically seeded
            value so goal-vs-actual reflects reality.
        rationale:
          type: string
    DraftBusinessCase:
      type: object
      required:
        - baseline
        - targetPayback
        - description
      properties:
        baseline:
          type: string
        targetPayback:
          type: string
        description:
          type: string
    AnalyzeSourceInput:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
        platform:
          type: string
        nameHint:
          type: string
...
    FetchSourceInput:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          minLength: 1
    FetchedFile:
      type: object
      required:
        - path
        - bytes
      properties:
        path:
          type: string
        bytes:
          type: integer
    FetchSourceResult:
      type: object
      required:
        - content
        - sourceType
        - files
        - truncated
      properties:
        content:
          type: string
        sourceType:
          type: string
          enum: [git, url]
        files:
          type: array
          items:
            $ref: "#/components/schemas/FetchedFile"
        truncated:
          type: boolean
    AgentDraft:
      type: object
      required:
        - name
        - role
        - tagline
        - bio
        - shouldDo
        - shouldNotDo
        - autonomyLevel
        - limits
        - businessCase
        - proposedMetrics
        - summary
        - confidence
      properties:
        name:
          type: string
        role:
          type: string
        tagline:
          type: string
        bio:
          type: string
        shouldDo:
          type: array
          items:
            type: string
        shouldNotDo:
          type: array
          items:
            type: string
        autonomyLevel:
          type: string
          enum: [autonomous, escalates, restricted]
        autonomyNotes:
          type: string
        limits:
          type: array
          items:
            type: string
        businessCase:
          $ref: "#/components/schemas/DraftBusinessCase"
        proposedMetrics:
          type: array
          items:
            $ref: "#/components/schemas/DraftMetric"
        summary:
          type: string
        confidence:
          type: number
...